requirements.txt
Django~=3.1.7
django-debug-toolbar~=3.2
django-mptt~=0.12.0